move JWKS plugin to `utilities/fastify`
